start consolidating all  app clients into app_client.py; reduces all of the separate module-level versions
